+++ src/main/java/com/example/team2_be/auth/KakaoAuthClient.java
@@ -1,7 +1,6 @@
 package com.example.team2_be.auth;
 
-import com.example.team2_be.auth.dto.kakao.KakaoAccessTokenRequestDTO;
-import com.example.team2_be.auth.dto.kakao.KakaoAccountDTO;
+import com.example.team2_be.auth.dto.kakao.KakaoInfoDTO;
 import com.example.team2_be.auth.dto.kakao.KakaoTokenDTO;
 import com.example.team2_be.core.config.AuthFeignConfig;
 import org.springframework.cloud.openfeign.FeignClient;
@@ -12,7 +11,7 @@ import java.net.URI;
 @FeignClient(name = "kakaoAuth", configuration = AuthFeignConfig.class)
 public interface KakaoAuthClient {
     @GetMapping
-    KakaoAccountDTO getInfo(URI baseUrl, @RequestHeader("Authorization") String accessToken);
+    KakaoInfoDTO getInfo(URI baseUrl, @RequestHeader("Authorization") String accessToken);
 
     @PostMapping
     KakaoTokenDTO getToken(URI baseUrl, @RequestParam("client_id") String restApiKey,
