+++ src/main/java/com/example/team2_be/auth/dto/UserAccountDTO.java
@@ -6,6 +6,11 @@ import lombok.ToString;
 @Getter
 @ToString
 public class UserAccountDTO {
-    protected String nickname;
     protected String email;
+    protected String nickname;
+
+    public UserAccountDTO(String email, String nickname) {
+        this.email = email;
+        this.nickname = nickname;
+    }
 }
