+++ src/main/java/com/example/team2_be/auth/dto/kakao/KakaoInfoDTO.java
@@ -1,4 +1,4 @@
-package com.example.team2_be.auth.dto.google;
+package com.example.team2_be.auth.dto.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -6,11 +6,10 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
-@Getter
 @ToString
+@Getter
 @NoArgsConstructor
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class GoogleAccountDTO{
-    private String email;
-    private String name;
+public class KakaoInfoDTO {
+    private KakaoAccountDTO kakaoAccount;
 }
